Introduce test program
Test program that will be used for corrections.
It works a bit differently than previous test program in that it attempts to run as many tests as possible (even if they may crash) and reports a result of the form M/N where M is the number of tests passed and N is the number of tests ran.
It also returns N - M as its return value.