Make TEST_FILE macro line-ending agnostic.
Since many students are developing on Windows, the code that produces db_modifiee.txt adds Windows-style line-endings to the file db_modifiee.txt. Then the test macro TEST_FILE compares that file with another that was produced in a Linux environment with Unix-style linde endings and the test fails. The fix is to use diff's switch to not compare line endings.